refactor of quests such that individual quest types are now possible
