+++ app/assets/v2/js/pages/quests.index.js
@@ -1,0 +1,42 @@
+
+$(document).ready(function() {
+
+  $('#reflink').click(function() {
+    $(this).focus();
+    $(this).select();
+    document.execCommand('copy');
+    $(this).after('<div class=after_copy>Copied to clipboard</div>');
+    setTimeout(function() {
+      $('.after_copy').remove();
+    }, 500);
+  });
+
+  $('.demo').click(function(e) {
+    e.preventDefault();
+    $(this).fadeOut(function() {
+      $('.demo').fadeIn();
+      var src = $('.demo').attr('src') + '?';
+      
+      $('.demo').attr('src', src);
+    });
+  });
+
+  $('#tabs a').click(function(e) {
+    e.preventDefault();
+    var target = $(this).data('href');
+
+    $('.difficulty_tab').addClass('hidden');
+    $('.nav-link').removeClass('active');
+    $(this).addClass('active');
+    $('.difficulty_tab.' + target).removeClass('hidden');
+  });
+
+  $('.quest-card.available').click(function(e) {
+    e.preventDefault();
+    document.location.href = $(this).find('a').attr('href');
+  });
+  $('.quest-card.available').mouseover(function(e) {
+    random_attn_effect($(this).find('.btn'));
+  });
+
+});
