+++ app/quests/templates/quests/base.html
@@ -44,5 +44,6 @@
   <script>
   </script>
   <script src="{% static "v2/js/pages/quests.helpers.js" %}"></script>
-  <script src="{% static "v2/js/pages/quests.js" %}"></script>
+  {% block 'scripts' %}
+  {% endblock %}
 </html>
