+++ app/quests/templates/quests/index.html
@@ -330,5 +330,11 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
   </div>
 </div>
 
+<script src="{% static "v2/js/pages/quests.index.js" %}"></script>
 
 {% endblock %}
+
+{% block 'scripts' %}
+<script src="{% static "v2/js/pages/quests.index.js" %}"></script>
+{% endblock %}
+
