+++ app/quests/templates/quests/quest.html
@@ -110,4 +110,8 @@
     };
     document.music_enabled = true;
   </script>
+  <script src="{% static "v2/js/pages/quests.quest.quiz_style.js" %}"></script>
+{% endblock %}
+{% block 'scripts' %}
+<script src="{% static "v2/js/pages/quests.quest.quiz_style.js" %}"></script>
 {% endblock %}
